# docs/design/openapi.yaml
openapi: 3.1.0
info:
  title: Multi-Agent AI Platform API
  version: "1.0.0"
  description: >
    Modular Monolith · Hexagonal · Plugin · Event-Driven.
    Auth: Firebase ID Token (local JWT verify). Tenant (workspace_id) derived from identity, never client-supplied.
    Errors: RFC 9457 Problem Details. Pagination: cursor-based. JSON: snake_case.
    All collection endpoints are wrapped as { data: [...], meta: { next_cursor, limit } } (API-04);
    bounded lists set next_cursor: null. Every response returns X-Correlation-Id.
    The interactive WebSocket endpoint /api/v1/ws is out of scope for OpenAPI and is specified in 03-api-spec.md §3.2.
servers:
  - url: /api/v1
security:
  - firebaseBearer: []
...
paths:
...
  /spaces:
    get:
      tags: [spaces]
      operationId: listSpaces
      description: "The workspace's active spaces, newest first, each with what it holds. `bytes_used` counts ACTIVE files only and is what the 1 GiB per-space quota is measured against."
      parameters: [ { $ref: "#/components/parameters/Limit" }, { $ref: "#/components/parameters/Cursor" } ]
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/SpacePage" } } } }
        default: { $ref: "#/components/responses/Problem" }
    post:
      tags: [spaces]
      operationId: createSpace
      description: "Mint a space. A name already taken in this workspace is `spaces.duplicate_name`/409, refused by the unique index in the same statement that would have written the row. The counters come back as zeros."
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/SpaceCreateIn" } } } }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/SpaceOut" } } } }
        default: { $ref: "#/components/responses/Problem" }
  /spaces/{id}:
    patch:
      tags: [spaces]
      operationId: renameSpace
      description: "Rename a space - the only mutable field it has. Unknown is 404 and soft-deleted is 409, the read/write asymmetry every other module uses. Renaming to the current name succeeds and writes nothing."
      parameters: [ { $ref: "#/components/parameters/Id" } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/SpaceRenameIn" } } } }
      responses:
        "200": { description: Renamed, content: { application/json: { schema: { $ref: "#/components/schemas/SpaceOut" } } } }
        default: { $ref: "#/components/responses/Problem" }
    delete:
      tags: [spaces]
      operationId: deleteSpace
      description: "Delete a space AND everything it owns: its documents and their vectors, its files and their stored objects, its threads and their messages and pins. There is no undo. Idempotent BY CONSTRUCTION rather than by an `Idempotency-Key` ledger - every step is a no-op when repeated, so retrying REPAIRS a cascade that died half-way, which a replayed response would instead skip."
      parameters: [ { $ref: "#/components/parameters/Id" } ]
      responses:
        "204": { description: Deleted with its contents }
        default: { $ref: "#/components/responses/Problem" }
...
  /workflows:
    get:
      tags: [workflows]
      operationId: listWorkflows
      responses:
        "200": { description: OK, headers: { X-Correlation-Id: { $ref: "#/components/headers/CorrelationId" } }, content: { application/json: { schema: { $ref: "#/components/schemas/WorkflowPage" } } } }
        default: { $ref: "#/components/responses/Problem" }
  /workflows/{key}/run:
    post:
      tags: [workflows]
      operationId: runWorkflow
      parameters: [ { name: key, in: path, required: true, schema: { type: string } }, { $ref: "#/components/parameters/CorrelationId" }, { $ref: "#/components/parameters/IdempotencyKey" } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/WorkflowRunIn" } } } }
      responses:
        "200":
          description: Result (JSON) or SSE stream
          content:
            application/json: { schema: { $ref: "#/components/schemas/WorkflowRunOut" } }
            text/event-stream: { schema: { type: string } }
        "202": { description: Accepted (heavy async), content: { application/json: { schema: { $ref: "#/components/schemas/WorkflowRunOut" } } } }
        default: { $ref: "#/components/responses/Problem" }
  /workflows/runs/{id}:
    get:
      tags: [workflows]
      operationId: getWorkflowRun
      parameters: [ { $ref: "#/components/parameters/Id" } ]
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/WorkflowRunOut" } } } }
        default: { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    firebaseBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: string, format: uuid } }
...
    Limit: { name: limit, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
    Cursor: { name: cursor, in: query, required: false, schema: { type: string } }
...
    CorrelationId: { name: X-Correlation-Id, in: header, required: false, schema: { type: string } }
    IdempotencyKey: { name: Idempotency-Key, in: header, required: false, schema: { type: string } }
  headers:
    CorrelationId:
      description: "Request correlation id, echoed on every response (generated if absent). See 03 §0."
      schema: { type: string }
  responses:
    Problem:
      description: RFC 9457 Problem Details
      headers:
        X-Correlation-Id: { $ref: "#/components/headers/CorrelationId" }
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    Problem:
      type: object
      required: [type, title, status, code, correlation_id]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code: { type: string, description: "stable machine code, e.g. files.too_large" }
        correlation_id: { type: string }
        errors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }
    PageMeta:
      type: object
      properties: { next_cursor: { type: string, nullable: true, description: "opaque cursor; null on the last (or unbounded) page" }, limit: { type: integer } }
...
    SpaceCreateIn:
      type: object
      required: [name]
      properties: { name: {type: string, minLength: 1, maxLength: 120} }
    SpaceRenameIn:
      type: object
      required: [name]
      properties: { name: {type: string, minLength: 1, maxLength: 120} }
    SpaceOut:
      type: object
      properties:
        id: {type: string, format: uuid}
        name: {type: string}
        bytes_used: {type: integer, description: "Total size of the space's ACTIVE files - a soft-deleted file has already given its bytes back to the quota."}
        file_count: {type: integer}
        conversation_count: {type: integer}
        created_at: {type: string, format: date-time, description: "UTC ISO-8601, trailing Z (DAT-06)", pattern: "Z$"}
    SpacePage:
      type: object
      properties: { data: { type: array, items: { $ref: "#/components/schemas/SpaceOut" } }, meta: { $ref: "#/components/schemas/PageMeta" } }
...
    WorkflowOut:
      type: object
      properties: { key: {type: string}, name: {type: string}, steps: { type: array, items: {type: string} } }
    WorkflowRunIn:
      type: object
      required: [space_id, input]
      properties: { space_id: {type: string, format: uuid, description: "The space this run's thread is opened in. Required with no nullable case: a run always opens its own thread, so it never has one to inherit a space from."}, input: {type: object, additionalProperties: true}, stream: {type: boolean, default: false} }
    WorkflowRunOut:
      type: object
      properties:
        run_id: {type: string, format: uuid}
        conversation_id: {type: string, format: uuid}
        status: {type: string, enum: [running, succeeded, failed]}
...
    WorkflowPage:
      type: object
      properties: { data: { type: array, items: { $ref: "#/components/schemas/WorkflowOut" } }, meta: { $ref: "#/components/schemas/PageMeta" } }
...